backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.1
python-decouple==3.8
redis==4.5.1
fastapi-cache2==0.2.1
aioredis==2.0.1
python-multipart==0.0.6
pandas==2.1.4
joblib==1.3.2
pymongo==4.6.1
pydantic==1.10.13
starlette==0.27.0
typing-extensions==4.9.0
numpy==1.26.3
scikit-learn==1.3.2
python-dateutil==2.8.2
pytz==2023.3
sendgrid==6.9.7
python-http-client==3.3.7
starkbank-ecdsa==2.2.0



